+++ config/routes.rb
@@ -280,6 +280,7 @@ Rails.application.routes.draw do
 
   namespace :api do
     namespace :v1 do
+      match 'mcp', to: 'mcp#handle', via: %i[get post delete]
       get   'photos', to: 'photos#index'
       get   'health', to: 'health#index'
       patch 'settings', to: 'settings#update'